pyproject.toml
[project]
name = "nft-weingarten-penalty-tracker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pandas>=2.3.2",
    "plotly-express>=0.4.1",
    "streamlit>=1.50.0",
    "streamlit-extras>=0.7.8",
    "numpy>=1.20.0",
    "black>=25.9.0",
]
